Add Channel for Response Handling
Added a channel to allow worker functions to give response output
to the HTTP handlers. I still need to figure out a better way
to organize the construction of these requests in the worker
functions, but currently the passing of responses out
of the worker function is functional on the Member worker.